pyproject.toml
[project]
name = "language-detective"
version = "0.1.0"
description = "A service that detects spoken language in audio files using multiple AI providers"
authors = [
    {name = "DripLink BE Intern", email = "intern@driplink.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]
